Share the pack-and-install smoke orchestration
The publish load smoke and the tool-load smoke both packed a set of
packages into a scratch registry and installed them into a fresh
consumer, so internal @intx/* deps resolve to the packed tarballs. A
shared step now does that; both scripts pass their package set and load
from the returned consumer. The tool-load smoke stops suppressing
install output with it, so a failing install surfaces its own diagnostic
there too.
